test(l1): move L2-only tx-type test to ethrex-test crate, clarify doc comment
