+++ test/tests/blockchain/mod.rs
@@ -1,5 +1,6 @@
 mod batch_tests;
 mod eip7702_revert_authority_tests;
 mod eip7702_zero_transfer_tests;
+mod l1_tx_type_tests;
 mod mempool_tests;
 mod smoke_tests;
